+++ docs/pages/docs/_sidebar.mdx
@@ -62,7 +62,6 @@ Functionality Plugins
 - [Conventional Commits](./generated/conventional-commits)
 - [Exec](./generated/exec)
 - [First Time Contributor](./generated/first-time-contributor)
-- [GitHub Action Toggle Peer Review](./generated/github-action-toggle-peer-review)
 - [GitHub Pages](./generated/gh-pages)
 - [Jira](./generated/jira)
 - [Omit Commits](./generated/omit-commits)
