--- plugins/gh-action-toggle-peer-review/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@auto-it/gh-action-toggle-peer-review",
-  "version": "9.51.0",
-  "main": "dist/index.js",
-  "description": "Toggle 'Require pull request reviews before merging' when creating 'latest' release from a GitHub Action",
-  "license": "MIT",
-  "author": {
-    "name": "Andrew Lisowski",
-    "email": "lisowski54@gmail.com"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/intuit/auto"
-  },
-  "files": [
-    "dist"
-  ],
-  "keywords": [
-    "automation",
-    "semantic",
-    "release",
-    "github",
-    "labels",
-    "automated",
-    "continuos integration",
-    "changelog"
-  ],
-  "scripts": {
-    "build": "tsc -b",
-    "start": "npm run build -- -w",
-    "lint": "eslint src --ext .ts",
-    "test": "jest --maxWorkers=2 --config ../../package.json"
-  },
-  "dependencies": {
-    "@auto-it/core": "link:../../packages/core",
-    "@octokit/rest": "^18.0.0",
-    "endent": "^2.0.1",
-    "tslib": "1.10.0"
-  }
-}
