--- plugins/gh-action-toggle-peer-review/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*", "../../typings/**/*"],
-
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "composite": true
-  },
-
-  "references": [
-    {
-      "path": "../../packages/core"
-    }
-  ]
-}
